Organize server.js: Extract schemas, models, config, and API routes into separate folders
