+++ routes/api.js
@@ -1,0 +1,108 @@
+const express = require('express');
+const { BuyOrder, BannedUser, generateOrderId } = require('../models');
+const { adminIds } = require('../config');
+
+const router = express.Router();
+
+// Wallet Address Endpoint
+router.get('/get-wallet-address', (req, res) => {
+    try {
+        const walletAddress = process.env.WALLET_ADDRESS;
+        
+        if (!walletAddress) {
+            return res.status(500).json({
+                success: false,
+                error: 'Wallet address not configured'
+            });
+        }
+
+        res.json({
+            success: true,
+            walletAddress: walletAddress
+        });
+    } catch (error) {
+        console.error('Error getting wallet address:', error);
+        res.status(500).json({
+            success: false,
+            error: 'Internal server error'
+        });
+    }
+});
+
+// Order Creation Endpoint
+router.post('/orders/create', async (req, res) => {
+    try {
+        const { telegramId, username, stars, walletAddress, isPremium, premiumDuration } = req.body;
+
+        if (!telegramId || !username || !walletAddress || (isPremium && !premiumDuration)) {
+            return res.status(400).json({ error: 'Missing required fields' });
+        }
+
+        const bannedUser = await BannedUser.findOne({ users: telegramId.toString() });
+        if (bannedUser) {
+            return res.status(403).json({ error: 'You are banned from placing orders' });
+        }
+
+        const priceMap = {
+            regular: { 1000: 20, 500: 10, 100: 2, 50: 1, 25: 0.6, 15: 0.35 },
+            premium: { 3: 19.31, 6: 26.25, 12: 44.79 }
+        };
+
+        let amount, packageType;
+        if (isPremium) {
+            packageType = 'premium';
+            amount = priceMap.premium[premiumDuration];
+        } else {
+            packageType = 'regular';
+            amount = priceMap.regular[stars];
+        }
+
+        if (!amount) {
+            return res.status(400).json({ error: 'Invalid selection' });
+        }
+
+        const order = new BuyOrder({
+            id: generateOrderId(),
+            telegramId,
+            username,
+            amount,
+            stars: isPremium ? null : stars,
+            premiumDuration: isPremium ? premiumDuration : null,
+            walletAddress,
+            isPremium,
+            status: 'pending',
+            dateCreated: new Date(),
+            adminMessages: []
+        });
+
+        await order.save();
+
+        const userMessage = isPremium ?
+            `🎉 Premium order received!\n\nOrder ID: ${order.id}\nAmount: ${amount} USDT\nDuration: ${premiumDuration} months\nStatus: Pending` :
+            `🎉 Order received!\n\nOrder ID: ${order.id}\nAmount: ${amount} USDT\nStars: ${stars}\nStatus: Pending`;
+
+        // Note: bot.sendMessage will be handled by the main server
+        // This is just the API endpoint logic
+
+        const adminMessage = isPremium ?
+            `🛒 New Premium Order!\n\nOrder ID: ${order.id}\nUser: @${username}\nAmount: ${amount} USDT\nDuration: ${premiumDuration} months` :
+            `🛒 New Buy Order!\n\nOrder ID: ${order.id}\nUser: @${username}\nAmount: ${amount} USDT\nStars: ${stars}`;
+
+        const adminKeyboard = {
+            inline_keyboard: [[
+                { text: '✅ Complete', callback_data: `complete_buy_${order.id}` },
+                { text: '❌ Decline', callback_data: `decline_buy_${order.id}` }
+            ]]
+        };
+
+        // Note: Admin notifications will be handled by the main server
+        // This is just the API endpoint logic
+
+        res.json({ success: true, order });
+    } catch (err) {
+        console.error('Order creation error:', err);
+        res.status(500).json({ error: 'Failed to create order' });
+    }
+});
+
+module.exports = router;
